requirements.txt
fastapi
duckdb
python-dotenv
openai
pandas
pydantic
PyYAML
networkx
